rjvbn

+++ LibraryProject.Data/Entity/BookAuthor.cs
@@ -2,10 +2,10 @@
 {
     public class BookAuthor
     {
-        public int BookId { get; set; }
+        public long BookId { get; set; }
         public Book Book { get; set; }
 
-        public int AuthorId { get; set; }
+        public long AuthorId { get; set; }
         public Author Author { get; set; }
     }
 }
