+++ LibraryProject.Data/Entity/BorrowBook.cs
@@ -8,10 +8,10 @@ namespace LibraryProject.Data.Entity
 {
     public class BorrowBook
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public long BorrowId { get; set; }
-        public int BookId { get; set; }
-        public int UserId { get; set; }
+        public long BookId { get; set; }
+        public long UserId { get; set; }
         public int Amount { get; set; }
         public DateTime DueDate { get; set; }
         public Borrow Borrow { get; set; }
